Pull request fixes
Rebase function in other file. Fix minor comments.
